+++ src/geo/geohash.clj
@@ -36,12 +36,18 @@
   GeoHash
   (to-shape [^GeoHash geohash] (bbox geohash))
   (to-jts ([^GeoHash geohash] (bbox-geom geohash))
-          ([^GeoHash geohash srid] (jts/set-srid (spatial/to-jts geohash) srid)))
+          ([^GeoHash geohash srid] (jts/set-srid (spatial/to-jts geohash) srid))
+          ([this c1 c2] (spatial/to-jts (spatial/to-jts this) c1 c2))
+          ([this c1 c2 geometry-factory] (spatial/to-jts (spatial/to-jts this)
+                                                         c1 c2 geometry-factory)))
 
   WGS84Point
   (to-shape [this] (spatial/spatial4j-point this))
   (to-jts ([this] (spatial/jts-point this))
-          ([this srid] (spatial/to-jts (spatial/to-jts this) srid))))
+          ([this srid] (spatial/to-jts (spatial/to-jts this) srid))
+          ([this c1 c2] (spatial/to-jts (spatial/to-jts this) c1 c2))
+          ([this c1 c2 geometry-factory] (spatial/to-jts (spatial/to-jts this)
+                                                         c1 c2 geometry-factory))))
 
 (defn northern-neighbor [^GeoHash h] (.getNorthernNeighbour h))
 (defn eastern-neighbor [^GeoHash h] (.getEasternNeighbour h))